requirements.txt
streamlit>=1.30.0
fpdf2>=2.7.0
Jinja2>=3.1.0
